requirements.txt
tensorflow
gradio
pillow
numpy<2